Moves internal MSK package to Kafka to match API package
