+++ aws/resource_aws_msk_cluster_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/terraform-providers/terraform-provider-aws/aws/internal/keyvaluetags"
-	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/msk/lister"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/kafka/lister"
 )
 
 func init() {
